Interactive Discovery

+++ application/views/app/discovery.php
@@ -140,14 +140,14 @@ echo view_i_nav(true, $focus_i);
 
 
         var $win = $(window);
-        if ((parseInt($("body").height())+89) <= parseInt($win.height())) {
+        if ((parseFloat($("body").height())+89) <= parseFloat($win.height())) {
             console.log('step1/'+$("body").height()+'/'+$win.height());
             $(".fixed-bottom").removeClass('hidden');
         } else {
             $(function () {
                 $win.scroll(function () {
                     //Download loading from bottom:
-                    if (parseInt($(document).height() - ($win.height() + $win.scrollTop())) < 144) {
+                    if (parseFloat($(document).height() - ($win.height() + $win.scrollTop())) < 144) {
                         console.log('step2');
                         $(".fixed-bottom").removeClass('hidden');
                     }
